requirements.txt
python-dotenv
Flask
pymongo[srv]
flask-restful
pyjwt
Flask-JWT-Extended